refactor: restructure models and remove unused task model; update user model with new fields
