+++ app/models.py
@@ -1,31 +1,35 @@
 from app.helpers.extensions import db
+from datetime import datetime
 
-
-# class User(db.Model):
-#     id = db.Column(db.Integer, primary_key=True)
-#     username = db.Column(db.String(80), unique=True, nullable=False)
-#     email = db.Column(db.String(120), unique=True, nullable=False)
-#     tasks = db.relationship("Task", backref="user", lazy=True)
-
-
-# class Task(db.Model):
-#     id = db.Column(db.Integer, primary_key=True)
-#     title = db.Column(db.String(255), nullable=False)
-#     description = db.Column(db.Text, nullable=True)
-#     completed = db.Column(db.Boolean, default=False)
-#     user_id = db.Column(db.Integer, db.ForeignKey("user.id"), nullable=False)
-
-class Project(db.Model):
+class User(db.Model):
     id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(80), unique=True, nullable=False)
     email = db.Column(db.String(120), unique=True, nullable=False)
-    tasks = db.relationship("Task", backref="user", lazy=True)
-    password_hash = db.Column(db.String(1024), nullable=False)
+    password_hash = db.Column(db.String(255), nullable=False)
+    username = db.Column(db.String(100), nullable=False)
+    two_factor_enabled = db.Column(db.Boolean, default=False)
+    created_at = db.Column(db.DateTime, default=datetime.utcnow)
+
+    # tasks_created = db.relationship("Task", foreign_keys='Task.created_by', backref="creator", lazy=True)
+    # tasks_assigned = db.relationship("Task", foreign_keys='Task.assignee_id', backref="assignee", lazy=True)
 
 
 class Task(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     title = db.Column(db.String(255), nullable=False)
     description = db.Column(db.Text, nullable=True)
-    completed = db.Column(db.Boolean, default=False)
-    user_id = db.Column(db.Integer, db.ForeignKey("user.id"), nullable=False)
+    status = db.Column(db.String(50), default="To Do")
+    est_time = db.Column(db.Float)
+    due_date = db.Column(db.DateTime)
+    priority = db.Column(db.String(20)) #low, medium, high
+    assignee_id = db.Column(db.Integer, db.ForeignKey("user.id"))
+    project_id = db.Column(db.Integer)
+    created_by = db.Column(db.Integer, db.ForeignKey("user.id"))
+    created_at = db.Column(db.DateTime, default=datetime.utcnow)
+    updated_at = db.Column(db.DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
+
+class Project(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(255), nullable=False)
+    description = db.Column(db.Text)
+    created_by = db.Column(db.Integer, db.ForeignKey("user.id"), nullable=False)
+    created_at = db.Column(db.DateTime, default=datetime.utcnow)
